arabizi_dataset_generator/requirements.txt
python-dotenv
pyyaml
openai>=1.0.0
tqdm>=4.66.1
python-dotenv>=1.0.0
pandas>=2.0.0